+++ react/components/EXPERIMENTAL_Table/index.tsx
@@ -5,7 +5,7 @@ import { TableProvider } from './context'
 import Toolbar from './Toolbar/index'
 
 import { DENSITY_OPTIONS } from './constants'
-import LineActions, { LineActionProps } from './LineActions/index'
+import LineActions, { LineActionProps } from './LineActions'
 import Pagination, { PaginationProps } from './Pagination'
 import { STATE_NOT_FOUND_ERROR } from './errors'
 import { TableContainer, Thead } from './Styled'
@@ -69,8 +69,8 @@ export type Props = InferProps<typeof tablePropTypes>
 
 export type TableComposites = {
   Toolbar: FC
-  Pagination: FC<PaginationProps>
-  LineActions: FC<LineActionProps>
+  Pagination?: FC<PaginationProps>
+  LineActions?: FC<LineActionProps>
 }
 
 Table.Toolbar = Toolbar
